FusionIO: implement a thin helper library around FusionIO's KV API
Because FusionIO's key/value store API wouldn't be very easy to expose
in Java, fio_kv_helper is a small library that wraps the calls to
FusionIO's vsldpexp library into an easier to use and understand
interface.

This commit also provides the means of building this helper library via
the Maven module 'helper'. FusionIO's SDK must of course be
installed on the machine attempting to build it to provide the vsldpexp
library and header files.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>