FusionIO: implement JNA library mapping for fio_kv_helper
This change introduces the FusionIOAPI class into the Java world,
providing methods to access the natively mapped fio_kv_helper C++
library to access the FusionIO key/value store.

The API is mapped to the native functions using JNA and its
direct-mapping feature and presents a simple Java-esque interface to the
user.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>