Create a fiokv-java distribution module that bundles the build outputs
Introduces a new Maven sub-module, fiokv-java-dist, that depends on the
api and helper modules and uses a custom Maven assembly to bring the
build outputs of all the fiokv-java-* artifacts together. It needs to be
a distinct module that depends on the first two so that Maven can get
the module build order right.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>