Export fio_kv_get_last_error() through JNI
Display the errno error code in the exception when opening the
FusionIOAPI fails. Since errno (and thus fio_kv_get_last_error()) is not
thread-safe, let's not use it anywhere else for error coding, but it's
helpful on open() to figure out if it's a permission problem, a path
problem, etc.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>