Improve fio_kv_exists() to make use of the key_info parameter
Since 0.4.0 the directKV API's nvm_kv_exists() function allows for a
nvm_kv_key_info_t structure pointer to be passed as an optional last
argument to retreive key/value pair information if the pair exists in
the key/value store.

This commit introduces the necessary changes to make use of that
functionality. At the Java API level, there is now two overloading
exists() methods, one with an extra KeyValueInfo argument and one
without.

In the helper library, a small hack is necessary to force this argument
being non-NULL because of a bug in the directKV API which doesn't
support a NULL argument as it should (and as it is documented in their
API documentation). This hack will be removed as soon as this bug is
fixed by FusionIO.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>