Major API rework for better pool support
This refactoring brings proper pool support to the FusionIO API by
making it more pool-oriented. Key/value operations are now done
logically on pools, while the store deals with open/close and pool
management.

Constructors and native operations are now declared as package private
as they shouldn't be exposed to users of the API directly.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>