Add more features to Java binding
Upgraded the Java binding and helper library to work with FusionIO's
0.7.0 release and implemented most of the missing APIs through the JNI
binding.

There also are a few modifications to the previously implemented methods
to either offer more flexibility, and in some situations simplify the
API calls.

Some general cleanup as well, split off the JNI bridge functions into a
distinct source file and normalized the asserts in all methods.

Signed-off-by: Maxime Petazzoni <mpetazzoni@turn.com>